Add tests for branch config

+++ test.sh
@@ -10,17 +10,25 @@
 #
 set -e
 rm -rf test
+
 echo 'clone test'
 node index.js example.json
 [ -d test/bigstraw ] || exit 1
 echo 'pull test'
 node index.js example.json
-echo 'branch clone test'
+echo 'branch clone test with -b'
 rm -rf test
 node index.js -b master example.json
 [ -d test/bigstraw ] || exit 1
-echo 'branch pull test'
+echo 'branch pull test with -b'
 node index.js -b master example.json
+rm -rf test
+echo 'branch clone in config'
+node index.js branch.json
+[ -e test/bigstraw/TESTINGFILE ] || exit 1
+echo 'branch pull in config'
+node index.js branch.json
+[ -e test/bigstraw/TESTINGFILE ] || exit 1
 rm -rf test
 echo 'clone with :: syntax test'
 node index.js coloncolon.json
@@ -28,14 +36,15 @@ node index.js coloncolon.json
 echo 'pull with :: syntax test'
 node index.js coloncolon.json
 rm -rf test
-echo 'branch clone with :: syntax test'
+echo 'branch clone with -b and :: syntax test'
 node index.js -b master coloncolon.json
 [ -d test/firebase ] || exit 1
-echo 'branch pull with :: syntax test'
+echo 'branch pull with -b and :: syntax test'
 node index.js -b master coloncolon.json
 echo 'test -s + --no-s'
 node index.js notssh.json -s --no-s
 [ -d test/polymer ] || exit 1
 REMOTE=$(GIT_DIR="test/polymer/.git" git config --get remote.origin.url)
 [ $REMOTE = "https://github.com/Polymer/polymer.git" ] || exit 1
+
 rm -rf test
